requirements.txt
praw
re
matplotlib
python-dotenv